+++ app/lib/general.lib.js
@@ -45,6 +45,36 @@ function calculateAndLogTimeTaken (startTime, message, data = {}) {
 }
 
 /**
+ * Determine the start and end date for the current financial year
+ *
+ * We often need to work out what the start and end date for the current financial year is. But because the financial
+ * year starts on 01-APR and finishes on 31-MAR what that year is will change dependent on the current date.
+ *
+ * @returns {Object} An object containing a `startDate` and `endDate`
+ */
+function currentFinancialYear () {
+  const currentDate = new Date()
+  const currentYear = currentDate.getFullYear()
+
+  let startYear
+  let endYear
+
+  // IMPORTANT! getMonth returns an integer (0-11). So, January is represented as 0 and December as 11. This is why
+  // we use 2 rather than 3 to refer to March
+  if (currentDate.getMonth() <= 2) {
+    // For example, if currentDate was 2022-02-15 it would fall in financial year 2021-04-01 to 2022-03-31
+    startYear = currentYear - 1
+    endYear = currentYear
+  } else {
+    // For example, if currentDate was 2022-06-15 it would fall in financial year 2022-04-01 to 2023-03-31
+    startYear = currentYear
+    endYear = currentYear + 1
+  }
+
+  return { startDate: new Date(startYear, 3, 1), endDate: new Date(endYear, 2, 31) }
+}
+
+/**
  * Returns the current time in nanoseconds. Used as part of logging how long something takes
  *
  * We often want to see how long a process takes and capture it in our logs. This can be especially useful when we
@@ -152,6 +182,7 @@ function timestampForPostgres () {
 
 module.exports = {
   calculateAndLogTimeTaken,
+  currentFinancialYear,
   currentTimeInNanoseconds,
   generateUUID,
   periodsOverlap,
